Remove unnecessary api context component, and move shared constants and fetchData() into config file
