+++ app/client/src/components/app.tsx
@@ -9,26 +9,24 @@ import {
 import "uswds/css/uswds.css";
 import "uswds/js/uswds.js";
 // ---
-import { serverBasePath } from "../index";
+import { serverBasePath, serverUrl, fetchData } from "../config";
 import Loading from "components/loading";
 import Login from "components/login";
 import Dashboard from "components/dashboard";
 import RebateForms from "routes/rebateForms";
 import RebateForm from "routes/rebateForm";
 import NotFound from "routes/notFound";
-import { useApiState, fetchData } from "contexts/api";
 import { useUserState, useUserDispatch } from "contexts/user";
 
 function ProtectedRoutes({ children }: { children: JSX.Element }) {
   const { pathname } = useLocation();
-  const { apiUrl } = useApiState();
   const { isAuthenticating, isAuthenticated } = useUserState();
   const dispatch = useUserDispatch();
 
   // check if user is already logged in or needs to be redirected to /login route
   useEffect(() => {
     dispatch({ type: "FETCH_USER_DATA_REQUEST" });
-    fetchData(`${apiUrl}/api/v1/user`)
+    fetchData(`${serverUrl}/api/v1/user`)
       .then((res) => {
         const { epaUserData, samUserData } = res;
         dispatch({ type: "USER_SIGN_IN" });
@@ -41,7 +39,7 @@ function ProtectedRoutes({ children }: { children: JSX.Element }) {
         dispatch({ type: "FETCH_USER_DATA_FAILURE" });
         dispatch({ type: "USER_SIGN_OUT" });
       });
-  }, [apiUrl, dispatch, pathname]);
+  }, [dispatch, pathname]);
 
   if (isAuthenticating) {
     return <Loading />;
