+++ app/client/src/routes/rebateForms.tsx
@@ -1,20 +1,19 @@
 import { useEffect } from "react";
 import { Link } from "react-router-dom";
 // ---
+import { serverUrl, fetchData } from "../config";
 import Loading from "components/loading";
 import Message from "components/message";
-import { useApiState, fetchData } from "contexts/api";
 import { useFormsState, useFormsDispatch } from "contexts/forms";
 
 export default function RebateForms() {
-  const { apiUrl } = useApiState();
   const { rebateFormSubmissions } = useFormsState();
   const dispatch = useFormsDispatch();
 
   useEffect(() => {
     dispatch({ type: "FETCH_REBATE_FORM_SUBMISSIONS_REQUEST" });
 
-    fetchData(`${apiUrl}/api/v1/rebate-form-submissions`)
+    fetchData(`${serverUrl}/api/v1/rebate-form-submissions`)
       .then((res) => {
         dispatch({
           type: "FETCH_REBATE_FORM_SUBMISSIONS_SUCCESS",
@@ -24,7 +23,7 @@ export default function RebateForms() {
       .catch((err) => {
         dispatch({ type: "FETCH_REBATE_FORM_SUBMISSIONS_FAILURE" });
       });
-  }, [apiUrl, dispatch]);
+  }, [dispatch]);
 
   if (rebateFormSubmissions.status === "idle") {
     return null;
@@ -39,8 +38,6 @@ export default function RebateForms() {
       <Message type="error" text="Error loading rebate form submissions." />
     );
   }
-
-  // return <Loading />;
 
   return (
     <table className="usa-table usa-table--borderless usa-table--striped width-full">
